Change behavior of name_is_expired to not error out
This change creates an implication that all records already exist, but by default they are expired.

It also simplifies the behavior when ANS v2 needs to look into ANS v1 to know whether a name is expired. Since we can't catch errors in Move, it is simpler to not unnecessarily error for a view function like name_is_expired